+++ backend/ContainerApp/Manager/Services/Clients/AccessorClient.cs
@@ -25,10 +25,7 @@ public class AccessorClient(
         try
         {
             var task = await _daprClient.InvokeMethodAsync<TaskModel?>(
-                HttpMethod.Get,
-                "accessor",
-                $"task/{id}"
-            );
+                HttpMethod.Get, "accessor", $"tasks-accessor/task/{id}");
             _logger.LogDebug("Received task {TaskId} from Accessor service", id);
             return task;
         }
@@ -93,7 +90,7 @@ public class AccessorClient(
         );
         try
         {
-            await _daprClient.InvokeMethodAsync(HttpMethod.Delete, "accessor", $"task/{id}");
+            await _daprClient.InvokeMethodAsync(HttpMethod.Delete, "accessor", $"tasks-accessor/task/{id}");
             _logger.LogDebug("Task {TaskId} deletion request sent to Accessor service", id);
 
             return true;
@@ -165,11 +162,7 @@ public class AccessorClient(
         try
         {
             var chats = await _daprClient.InvokeMethodAsync<List<ChatSummary>>(
-                HttpMethod.Get,
-                "accessor",
-                $"chats/{userId}",
-                cancellationToken: ct
-            );
+                HttpMethod.Get, "accessor", $"chats-accessor/{userId}", cancellationToken: ct);
 
             return chats ?? new List<ChatSummary>();
         }
@@ -195,9 +188,7 @@ public class AccessorClient(
         try
         {
             return await _daprClient.InvokeMethodAsync<UserModel?>(
-                HttpMethod.Get,
-                "accessor",
-                $"users/{userId}");
+                HttpMethod.Get, "accessor", $"users-accessor/{userId}");
         }
         catch (Exception ex)
         {
@@ -210,11 +201,8 @@ public class AccessorClient(
     {
         try
         {
-            await _daprClient.InvokeMethodAsync(
-                HttpMethod.Post,
-                "accessor",
-                "users",
-                user);
+            await _daprClient.InvokeMethodAsync(HttpMethod.Post, "accessor", "users-accessor", user);
+
             return true;
         }
         catch (Exception ex)
@@ -228,11 +216,7 @@ public class AccessorClient(
     {
         try
         {
-            await _daprClient.InvokeMethodAsync(
-                HttpMethod.Put,
-                "accessor",
-                $"users/{userId}",
-                user);
+            await _daprClient.InvokeMethodAsync(HttpMethod.Put, "accessor", $"users-accessor/{userId}", user);
             return true;
         }
         catch (Exception ex)
@@ -246,10 +230,7 @@ public class AccessorClient(
     {
         try
         {
-            await _daprClient.InvokeMethodAsync(
-                HttpMethod.Delete,
-                "accessor",
-                $"users/{userId}");
+            await _daprClient.InvokeMethodAsync(HttpMethod.Delete, "accessor", $"users-accessor/{userId}");
             return true;
         }
         catch (Exception ex)
@@ -266,11 +247,7 @@ public class AccessorClient(
         try
         {
             var users = await _daprClient.InvokeMethodAsync<List<UserData>>(
-                HttpMethod.Get,
-                "accessor",
-                "users",
-                ct
-            );
+                HttpMethod.Get, "accessor", "users-accessor", ct);
 
             _logger.LogInformation("Retrieved {Count} users from accessor", users?.Count ?? 0);
             return users ?? Enumerable.Empty<UserData>();
@@ -287,11 +264,7 @@ public class AccessorClient(
         try
         {
             var snapshot = await _daprClient.InvokeMethodAsync<StatsSnapshot>(
-                HttpMethod.Get,
-                AppIds.Accessor,                 // keep using your constant if you have it
-                "internal/stats/snapshot",
-                ct
-            );
+                HttpMethod.Get, AppIds.Accessor, "internal-accessor/stats/snapshot", ct);
             return snapshot; // may be null if Accessor returns empty body
         }
         catch (InvocationException ex) when (
